fix(important): changed type of entities' ids to BigInteger and rewrote the tests, dto-s, and repositories

+++ src/main/java/com/epam/dto/TrainingDto.java
@@ -3,12 +3,13 @@ package com.epam.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 @Data
 @AllArgsConstructor
 public class TrainingDto {
-    private Long id;
+    private BigInteger id;
     private TraineeDto trainee;
     private TrainerDto trainer;
