+++ src/main/java/com/epam/repository/TrainingTypeRepository.java
@@ -3,6 +3,8 @@ package com.epam.repository;
 import com.epam.entity.TrainingTypeEntity;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public class TrainingTypeRepository extends MapRepositoryImpl<TrainingTypeEntity, Long> {
+public class TrainingTypeRepository extends MapRepositoryImpl<TrainingTypeEntity, BigInteger> {
 }
